feat: Implement configurable action pipelines
This commit introduces a new action engine and supporting infrastructure
to handle actions within event pipelines. This changes the format of the
configuration and deprecates the `internal/router` package.

Key changes:

- Implemented a new `pkg/pipeline` package to define core types for
action execution. - Implemented an action registry which enables the
actions to be defined using a factory pattern. Core actions are defined.
- Introduced a `config/compiler` package. - The `config.Load()` function
now uses the compiler to construct action pipelines from the loaded
configuration. - The `EventRouter` now executes pipelines.